requirements.txt
lxml==5.3.0
matplotlib==3.10.6
networkx==3.4.1
Pillow==11.3.0
PyQt5==5.15.11
PyQt5_sip==12.13.0
pyqtgraph==0.13.7
